backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
python-dotenv==1.0.0
cryptography==42.0.5
gunicorn==21.2.0
Werkzeug==3.0.1
